+++ model/geograficos/Departamento.js
@@ -37,7 +37,7 @@ module.exports.getAll = function(params) {
   let query = table.select(table.id, table.nombre)
   .from(table.join(Provincia.table).on(table.provincia.equals(Provincia.table.id)));
 
-  if (params.provincia_id) query.where(table.provincia.equals(params.provincia_id));
+  if (params.provincia_id && !isNaN(+params.provincia_id)) query.where(table.provincia.equals(+params.provincia_id));
   if (params.provincia_text) query.where(Provincia.table.nombre.equals(params.provincia_text));
 
   return new Promise(function(resolve, reject) {