+++ utils/index.js
@@ -11,3 +11,8 @@ module.exports.getFloat = function(value) {
   if (typeof value == 'number') return value;
   return value && value.length ? value.replace(',', '.') : null;
 }
+
+module.exports.numberOrNull = function(value) {
+  if (isNaN(+value)) return null;
+  else return +value;
+}